* elastic/master: (77 commits)
  Update ingest jdocs that a null return value will drop the current document. (#40359)
  Remove -Xlint exclusions in the ingest-common module. (#40505)
  Update docs for the DFR similarity (#40579)
  Fix merging of search_as_you_type field mapper (#40593)
  Support roles with application privileges against wildcard applications (#40398)
  Remove obsolete security settings (#40496)
  Remove Gradle deprecation warnings (#40449)
  Run the build integ test in parallel (#39788)
  Fix 3rd pary S3 tests (#40588)
  lower bwc skip for search as you type (#40599)
  Muting XContentParserTests#testSubParserArray
  Fixing typo in test error message (#40611)
  Update max dims for vectors to 1024. (#40597)
  Add start and stop time to cat recovery API (#40378)
  [DOCS] Correct keystore commands for Email and Jira actions in Watcher (#40417)
  [DOCS] Document common settings for snapshot repository plugins (#40475)
  Remove with(out)-system-key tests (#40547)
  Geo Point parse error fix (#40447)
  Handle null retention leases in WaitForNoFollowersStep (#40477)
  [DOCS] Adds anchors for ruby client (#39867)
  ...

+++ buildSrc/src/main/java/org/elasticsearch/gradle/testclusters/ElasticsearchNode.java
@@ -26,8 +26,10 @@ import org.gradle.api.logging.Logger;
 import org.gradle.api.logging.Logging;
 
 import java.io.BufferedReader;
+import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.UncheckedIOException;
 import java.net.HttpURLConnection;
@@ -39,15 +41,18 @@ import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Set;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Predicate;
+import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -71,6 +76,10 @@ public class ElasticsearchNode {
 
     private final LinkedHashMap<String, Predicate<ElasticsearchNode>> waitConditions;
     private final List<URI> plugins = new ArrayList<>();
+    private final Map<String, Supplier<CharSequence>> settings = new LinkedHashMap<>();
+    private final Map<String, Supplier<CharSequence>> keystoreSettings = new LinkedHashMap<>();
+    private final Map<String, Supplier<CharSequence>> systemProperties = new LinkedHashMap<>();
+    private final Map<String, Supplier<CharSequence>> environment = new LinkedHashMap<>();
 
     private final Path confPathRepo;
     private final Path configFile;
@@ -143,6 +152,55 @@ public class ElasticsearchNode {
         plugin(plugin.toURI());
     }
 
+    public void keystore(String key, String value) {
+        addSupplier("Keystore", keystoreSettings, key, value);
+    }
+
+    public void keystore(String key, Supplier<CharSequence> valueSupplier) {
+        addSupplier("Keystore", keystoreSettings, key, valueSupplier);
+    }
+
+    public void setting(String key, String value) {
+        addSupplier("Settings", settings, key, value);
+    }
+
+    public void setting(String key, Supplier<CharSequence> valueSupplier) {
+        addSupplier("Setting", settings, key, valueSupplier);
+    }
+
+    public void systemProperty(String key, String value) {
+        addSupplier("Java System property", systemProperties, key, value);
+    }
+
+    public void systemProperty(String key, Supplier<CharSequence> valueSupplier) {
+        addSupplier("Java System property", systemProperties, key, valueSupplier);
+    }
+
+    public void environment(String key, String value) {
+        addSupplier("Environment variable", environment, key, value);
+    }
+
+    public void environment(String key, Supplier<CharSequence> valueSupplier) {
+        addSupplier("Environment variable", environment, key, valueSupplier);
+    }
+
+    private void addSupplier(String name, Map<String, Supplier<CharSequence>> collector, String key, Supplier<CharSequence> valueSupplier) {
+        requireNonNull(key, name + " key was null when configuring test cluster `" + this + "`");
+        requireNonNull(valueSupplier, name + " value supplier was null when configuring test cluster `" + this + "`");
+        collector.put(key, valueSupplier);
+    }
+
+    private void addSupplier(String name, Map<String, Supplier<CharSequence>> collector, String key, String actualValue) {
+        requireNonNull(actualValue, name + " value was null when configuring test cluster `" + this + "`");
+        addSupplier(name, collector, key, () -> actualValue);
+    }
+
+    private void checkSuppliers(String name, Map<String, Supplier<CharSequence>> collector) {
+        collector.forEach((key, value) -> {
+            requireNonNull(value.get().toString(), name + " supplied value was null when configuring test cluster `" + this + "`");
+        });
+    }
+
     public Path getConfigDir() {
         return configFile.getParent();
     }
@@ -167,6 +225,8 @@ public class ElasticsearchNode {
     public File getJavaHome() {
         return javaHome;
     }
+
+
 
     private void waitForUri(String description, String uri) {
         waitConditions.put(description, (node) -> {
@@ -222,46 +282,79 @@ public class ElasticsearchNode {
             "install", "--batch", plugin.toString())
         );
 
+        if (keystoreSettings.isEmpty() == false) {
+            checkSuppliers("Keystore", keystoreSettings);
+            runElaticsearchBinScript("elasticsearch-keystore", "create");
+            keystoreSettings.forEach((key, value) -> {
+                runElaticsearchBinScriptWithInput(value.get().toString(), "elasticsearch-keystore", "add", "-x", key);
+            });
+        }
+
         startElasticsearchProcess();
     }
 
+    private void runElaticsearchBinScriptWithInput(String input, String tool, String... args) {
+        try (InputStream byteArrayInputStream = new ByteArrayInputStream(input.getBytes(StandardCharsets.UTF_8))) {
+            services.loggedExec(spec -> {
+                spec.setEnvironment(getESEnvironment());
+                spec.workingDir(workingDir);
+                spec.executable(
+                    OS.conditionalString()
+                        .onUnix(() -> "./bin/" + tool)
+                        .onWindows(() -> "cmd")
+                        .supply()
+                );
+                spec.args(
+                    OS.<List<String>>conditional()
+                        .onWindows(() -> {
+                            ArrayList<String> result = new ArrayList<>();
+                            result.add("/c");
+                            result.add("bin\\" + tool + ".bat");
+                            for (String arg : args) {
+                                result.add(arg);
+                            }
+                            return result;
+                        })
+                        .onUnix(() -> Arrays.asList(args))
+                        .supply()
+                );
+                spec.setStandardInput(byteArrayInputStream);
+
+            });
+        } catch (IOException e) {
+            throw new UncheckedIOException(e);
+        }
+    }
+
     private void runElaticsearchBinScript(String tool, String... args) {
-        services.loggedExec(spec -> {
-            spec.setEnvironment(getESEnvironment());
-            spec.workingDir(workingDir);
-            spec.executable(
-                OS.conditionalString()
-                    .onUnix(() -> "./bin/" + tool)
-                    .onWindows(() -> "cmd")
-                    .supply()
-            );
-            spec.args(
-                OS.<List<String>>conditional()
-                    .onWindows(() -> {
-                        ArrayList<String> result = new ArrayList<>();
-                        result.add("/c");
-                        result.add("bin\\" + tool + ".bat");
-                        for (String arg : args) {
-                            result.add(arg);
-                        }
-                        return result;
-                    })
-                    .onUnix(() -> Arrays.asList(args))
-                    .supply()
-            );
-        });
+        runElaticsearchBinScriptWithInput("", tool, args);
     }
 
     private Map<String, String> getESEnvironment() {
-        Map<String, String> environment= new HashMap<>();
-        environment.put("JAVA_HOME", getJavaHome().getAbsolutePath());
-        environment.put("ES_PATH_CONF", configFile.getParent().toString());
-        environment.put("ES_JAVA_OPTS", "-Xms512m -Xmx512m");
-        environment.put("ES_TMPDIR", tmpDir.toString());
+        Map<String, String> defaultEnv = new HashMap<>();
+        defaultEnv.put("JAVA_HOME", getJavaHome().getAbsolutePath());
+        defaultEnv.put("ES_PATH_CONF", configFile.getParent().toString());
+        String systemPropertiesString = "";
+        if (systemProperties.isEmpty() == false) {
+            checkSuppliers("Java System property", systemProperties);
+            systemPropertiesString = " " + systemProperties.entrySet().stream()
+                .map(entry -> "-D" + entry.getKey() + "=" + entry.getValue().get())
+                .collect(Collectors.joining(" "));
+        }
+        defaultEnv.put("ES_JAVA_OPTS", "-Xms512m -Xmx512m -ea -esa" + systemPropertiesString);
+        defaultEnv.put("ES_TMPDIR", tmpDir.toString());
         // Windows requires this as it defaults to `c:\windows` despite ES_TMPDIR
+        defaultEnv.put("TMP", tmpDir.toString());
 
-        environment.put("TMP", tmpDir.toString());
-        return environment;
+        Set<String> commonKeys = new HashSet<>(environment.keySet());
+        commonKeys.retainAll(defaultEnv.keySet());
+        if (commonKeys.isEmpty() == false) {
+            throw new IllegalStateException("testcluster does not allow setting the following env vars " + commonKeys);
+        }
+
+        checkSuppliers("Environment variable", environment);
+        environment.forEach((key, value) -> defaultEnv.put(key, value.get().toString()));
+        return defaultEnv;
     }
 
     private void startElasticsearchProcess() {
@@ -445,37 +538,49 @@ public class ElasticsearchNode {
     }
 
     private void createConfiguration()  {
-        LinkedHashMap<String, String> config = new LinkedHashMap<>();
+        LinkedHashMap<String, String> defaultConfig = new LinkedHashMap<>();
 
         String nodeName = safeName(name);
-        config.put("cluster.name",nodeName);
-        config.put("node.name", nodeName);
-        config.put("path.repo", confPathRepo.toAbsolutePath().toString());
-        config.put("path.data", confPathData.toAbsolutePath().toString());
-        config.put("path.logs", confPathLogs.toAbsolutePath().toString());
-        config.put("path.shared_data", workingDir.resolve("sharedData").toString());
-        config.put("node.attr.testattr", "test");
-        config.put("node.portsfile", "true");
-        config.put("http.port", "0");
-        config.put("transport.tcp.port", "0");
+        defaultConfig.put("cluster.name",nodeName);
+        defaultConfig.put("node.name", nodeName);
+        defaultConfig.put("path.repo", confPathRepo.toAbsolutePath().toString());
+        defaultConfig.put("path.data", confPathData.toAbsolutePath().toString());
+        defaultConfig.put("path.logs", confPathLogs.toAbsolutePath().toString());
+        defaultConfig.put("path.shared_data", workingDir.resolve("sharedData").toString());
+        defaultConfig.put("node.attr.testattr", "test");
+        defaultConfig.put("node.portsfile", "true");
+        defaultConfig.put("http.port", "0");
+        defaultConfig.put("transport.tcp.port", "0");
         // Default the watermarks to absurdly low to prevent the tests from failing on nodes without enough disk space
-        config.put("cluster.routing.allocation.disk.watermark.low", "1b");
-        config.put("cluster.routing.allocation.disk.watermark.high", "1b");
+        defaultConfig.put("cluster.routing.allocation.disk.watermark.low", "1b");
+        defaultConfig.put("cluster.routing.allocation.disk.watermark.high", "1b");
         // increase script compilation limit since tests can rapid-fire script compilations
-        config.put("script.max_compilations_rate", "2048/1m");
+        defaultConfig.put("script.max_compilations_rate", "2048/1m");
         if (Version.fromString(version).getMajor() >= 6) {
-            config.put("cluster.routing.allocation.disk.watermark.flood_stage", "1b");
+            defaultConfig.put("cluster.routing.allocation.disk.watermark.flood_stage", "1b");
         }
         if (Version.fromString(version).getMajor() >= 7) {
-            config.put("cluster.initial_master_nodes", "[" + nodeName + "]");
+            defaultConfig.put("cluster.initial_master_nodes", "[" + nodeName + "]");
         }
+        checkSuppliers("Settings", settings);
+        Map<String, String> userConfig = settings.entrySet().stream()
+            .collect(Collectors.toMap(entry -> entry.getKey(), entry -> entry.getValue().get().toString()));
+        HashSet<String> overriden = new HashSet<>(defaultConfig.keySet());
+        overriden.retainAll(userConfig.keySet());
+        if (overriden.isEmpty() ==false) {
+            throw new IllegalArgumentException("Testclusters does not allow the following settings to be changed:" + overriden);
+        }
+
         try {
             // We create hard links  for the distribution, so we need to remove the config file before writing it
             // to prevent the changes to reflect across all copies.
             Files.delete(configFile);
             Files.write(
                 configFile,
-                config.entrySet().stream()
+                Stream.concat(
+                    userConfig.entrySet().stream(),
+                    defaultConfig.entrySet().stream()
+                )
                     .map(entry -> entry.getKey() + ": " + entry.getValue())
                     .collect(Collectors.joining("\n"))
                     .getBytes(StandardCharsets.UTF_8)
